+++ assignment_login/src/App.jsx
@@ -1,41 +1,7 @@
 // eslint-disable-next-line no-unused-vars
 import React from "react";
-import { createBrowserRouter, RouterProvider, Outlet } from "react-router-dom";
-import Home from "./pages/Home";
-import Signup from "./pages/Signup";
-import Login from "./pages/Login";
-import Navbar from "./components/Navbar";
-
-const DashBoard = () => {
-  return (
-    <div>
-      <Navbar />
-      <Outlet />
-    </div>
-  );
-};
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <DashBoard />,
-    children: [
-      {
-        path: "/",
-        element: <Home />,
-      },
-    ],
-  },
-
-  {
-    path: "/Signup",
-    element: <Signup />,
-  },
-  {
-    path: "/login",
-    element: <Login />,
-  },
-]);
+import { RouterProvider } from "react-router-dom";
+import router from "./components/Common/Router";
 
 function App() {
   return (
