+++ assignment_login/src/components/Common/Router.jsx
@@ -1,0 +1,40 @@
+// eslint-disable-next-line no-unused-vars
+import React from "react";
+import { createBrowserRouter, Outlet } from "react-router-dom";
+import Home from "../../pages/Home";
+import Signup from "../../pages/Signup";
+import Login from "../../pages/Login";
+import Navbar from "../../components/Navbar";
+
+const DashBoard = () => {
+  return (
+    <div>
+      <Navbar />
+      <Outlet />
+    </div>
+  );
+};
+
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <DashBoard />,
+    children: [
+      {
+        path: "/",
+        element: <Home />,
+      },
+    ],
+  },
+
+  {
+    path: "/Signup",
+    element: <Signup />,
+  },
+  {
+    path: "/login",
+    element: <Login />,
+  },
+]);
+
+export default router;
